fix: extract color variables to independent file, and change some naing issues
