+++ src/components/DesignSystem/design-system.module.scss
@@ -1,21 +1,4 @@
-/* Primary Color*/
-$cyanDark: #2b788b;
-$cyanLight: #c3dce3;
-$pinkDark: #945069;
-$pinkLight: #f2d4dc;
-$black: #000000;
-$greyLight: #f6f5f4;
-$greyMedium: #e0e0e0;
-$greyIcon: #bababa;
-$greyDark: #757575;
-
-/* Secondary Color*/
-$cyan: #5996a5;
-$green: #639b6d;
-$pink: #a15993;
-$red: #a95151;
-$yellow: #c4a24c;
-$orange: #cb5b43;
+@import './color-theme.module.scss';
 
 @mixin headingBaseStyle {
   font-family: 'Dela Gothic One';
