+++ src/pages/design-system.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import DesignSystem from '../components/DesignSytem';
+import DesignSystem from '../components/DesignSystem';
 
 export default function DesignPage() {
   return <DesignSystem />;
